handleClick is not working
