+++ client/src/App.js
@@ -16,7 +16,7 @@ import PrivateRoute from "./components/private-route/PrivateRoute";
 import Dashboard from "./components/dashboard/Dashboard";
 import TalkRatio from './components/visualizations/TalkRatio';
 import Transcript from './components/visualizations/Transcript';
-import TurnTaking from './components/visualizations/TurnTaking';
+import TurnTaking from './components/visualizations/turn-taking/TurnTaking';
 
 // Check for token to keep user logged in
 if (localStorage.jwtToken) {
